requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2023.12.2
gprof2dot==2022.7.29
graphviz==0.20.1
huggingface-hub==0.20.3
idna==3.6
ijson==3.2.3
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pycocotools==2.0.7
pyjstat==2.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
six==1.16.0
snakeviz==2.2.0
sympy==1.12
timm==0.9.12
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0